refactor: 이력서 작성페이지 useEffect 내부 콜백함수 로직 분리
