+++ README.md
@@ -23,6 +23,7 @@
 + `TypeScript` : 정적 타입 검사로 더욱 안전한 개발환경
 + `Supabase` : 간단한 세팅으로 관계형데이터베이스 사용
 + `Tailwind CSS` : 우수한 성능과 CSS유지보수 상승
+![기술선택](https://github.com/user-attachments/assets/ee36080d-be12-4bb0-a50d-ce2694434809)
 
 ## 2. 프로젝트 구조
 ```
@@ -198,14 +199,17 @@ src
   표시됩니다.
 + 편집 버튼을 누르면 회원 정보를 편집할 수 있습니다.
 
-### [이력서, 이력서 추가]
-+ 멘티에게만 나의 이력서 목록이 표시됩니다.
-+ 내가 적은 이력서 목록이 표시되고 이력서를 추가할 수 있는 버튼이 나타납니다.
-+ 이력서 추가 버튼을 누르면 이력서를 적을 수 있는 폼이 나타납니다. 채택 포인트, 경력,
-  이름, 성별 등을 입력할 수 있습니다.
-+ 미리보기를 클릭하면 pdf 창이 뜨고 다운로드, 인쇄가 가능합니다.
-+ 이력서 등록을 클릭하면 이력서 목록에 이력서가 추가됩니다.
+### [이력서 상세]
++ 내가 적은 이력서가 화면에 출력되어 확인할 수 있습니다.
 + 멘토가 피드백을 달면 채택하여 채택 포인트를 멘토에게 전달할 수 있습니다.
++ 수정하기 버튼을 눌러 이력서 수정페이지로 이동할 수 있습니다.
+
+### [이력서 추가/수정]
++ 이름, 성별, 학력, 경력 등 다양한 정보를 입력하여 이력서 데이터를 채울 수 있습니다. 
++ 미리보기로 현재 이력서가 어떤 모습인지 확인할 수 있습니다.
++ 이력서 다운로드로 내 로컬환경에 이력서를 pdf로 저장할 수 있습니다.
++ 이력서 저장으로 DB에 내 이력서 파일을 업로드하고, 게시글을 생성하여 피드백을 받을 수 있습니다.
++ 이력서 수정기능으로 진입할 시 이전에 작성해두었던 데이터가 입력되어있어 원하는 데이터만 수정할 수 있습니다.
 
 ### [이력서 게시글 목록]
 + 멘토에게만 이력서 게시글 목록이 표시됩니다.
