+++ src/app/layout.tsx
@@ -16,8 +16,8 @@ const geistMono = localFont({
 });
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app'
+  title: '가취뽀',
+  description: '기술면접 연습과 이력서 피드백을 받으며 취업을 준비해요!'
 };
 
 export type Props = { points: 0; setPoints: (value: number) => void };
